# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: BilibiliComment Admin API
  version: 0.1.0
  description: 单管理员任务管理、评论观测与系统状态 API。所有 bigint ID 均按字符串传输。
servers:
  - url: /
security:
  - sessionCookie: []
...
paths:
...
  /api/v1/tasks/{taskId}/comments:
    parameters:
      - {$ref: '#/components/parameters/TaskId'}
    get:
      tags: [Executions]
      operationId: listTaskComments
      parameters:
        - {$ref: '#/components/parameters/Cursor'}
        - {$ref: '#/components/parameters/Limit'}
      responses:
        '200':
          description: 当前任务独立评论表的游标分页
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CommentPage'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/Unprocessable'}
...
  /api/v1/tasks/{taskId}/comment-exports:
    parameters:
      - {$ref: '#/components/parameters/TaskId'}
    post:
      tags: [Comments]
      operationId: createCommentExport
      security:
        - sessionCookie: []
          csrfHeader: []
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateCommentExportRequest'}
      responses:
        '202':
          description: 已创建异步导出任务；相同幂等请求返回原任务
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CommentExportJob'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '503': {$ref: '#/components/responses/Unavailable'}
    get:
      tags: [Comments]
      operationId: listCommentExports
      parameters:
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 20}
      responses:
        '200':
          description: 当前任务最近的评论导出任务
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/CommentExportJob'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/Unprocessable'}
  /api/v1/comment-exports/{exportId}:
    parameters:
      - {$ref: '#/components/parameters/ExportId'}
    get:
      tags: [Comments]
      operationId: getCommentExport
      responses:
        '200':
          description: 评论导出任务状态
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CommentExportJob'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      tags: [Comments]
      operationId: cancelCommentExport
      security:
        - sessionCookie: []
          csrfHeader: []
      responses:
        '204': {description: 已请求取消；终态任务保持不变}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '410': {$ref: '#/components/responses/Gone'}
        '422': {$ref: '#/components/responses/Unprocessable'}
  /api/v1/comment-exports/{exportId}/download:
    parameters:
      - {$ref: '#/components/parameters/ExportId'}
    get:
      tags: [Comments]
      operationId: downloadCommentExport
      responses:
        '200':
          description: 解密后的 CSV 或 JSONL 文件；响应禁止缓存
          headers:
            Content-Disposition:
              schema: {type: string}
            Cache-Control:
              schema: {type: string}
          content:
            text/csv:
              schema: {type: string, format: binary}
            application/x-ndjson:
              schema: {type: string, format: binary}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '410': {$ref: '#/components/responses/Gone'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/v1/credentials:
    get:
      tags: [Credentials]
      operationId: listCredentials
      responses:
        '200':
          description: 脱敏凭据列表
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/CredentialProfile'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/credentials/{credentialId}/secret:
    put:
      tags: [Credentials]
      operationId: replaceCredentialSecret
      security:
        - sessionCookie: []
          csrfHeader: []
      parameters:
        - {$ref: '#/components/parameters/CredentialId'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ReplaceCredentialSecretRequest'}
      responses:
        '204': {description: 已保存 Cookie 原文，响应不返回 Cookie 内容}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/Unprocessable'}
  /api/v1/credentials/{credentialId}/validate:
    post:
      tags: [Credentials]
      operationId: validateCredential
      security:
        - sessionCookie: []
          csrfHeader: []
      parameters:
        - {$ref: '#/components/parameters/CredentialId'}
      responses:
        '200':
          description: 验证后的凭据状态
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CredentialProfile'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '502': {$ref: '#/components/responses/UpstreamFailure'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
    csrfHeader:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
  parameters:
    TaskId:
      name: taskId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/PositiveId'}
...
    CredentialId:
      name: credentialId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/PositiveId'}
    ExportId:
      name: exportId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/PositiveId'}
    Cursor:
      name: cursor
      in: query
      description: 服务端返回的不透明分页边界；客户端必须原样回传，不得解析或自行构造。
      schema: {type: string, minLength: 1}
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 50}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 128}
  responses:
    BadRequest:
      description: 请求格式错误
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
    Unauthorized:
      description: 未登录或凭据错误
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
    Forbidden:
      description: CSRF 失效或来源域名被禁止
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
    NotFound:
      description: 资源不存在
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
    Conflict:
      description: 重复任务、活动执行或幂等冲突
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
    Gone:
      description: 导出任务或文件已过期并被清理
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
    Unprocessable:
      description: 参数验证失败
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
    RateLimited:
      description: 请求过于频繁
      headers:
        Retry-After: {schema: {type: integer, minimum: 1}}
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
    UpstreamFailure:
      description: Bilibili 上游响应异常
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
    Unavailable:
      description: 数据库或上游暂不可用
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
  schemas:
    PositiveId: {type: string, pattern: '^[1-9][0-9]*$'}
...
    CommentReplyScope: {type: string, enum: [ALL, ROOT, REPLIES], default: ALL}
    CommentSort: {type: string, enum: [CTIME_DESC, CTIME_ASC], default: CTIME_DESC}
    CommentExportFormat: {type: string, enum: [CSV, JSONL]}
    CommentExportColumn:
      type: string
      enum: [RPID, PARENT_RPID, MID, UNAME, AVATAR, CURRENT_LEVEL, CONTENT, CTIME]
    CommentExportStatus:
      type: string
      enum: [QUEUED, RUNNING, SUCCEEDED, FAILED, CANCELLED, EXPIRED]
...
    Comment:
      type: object
      required: [id, rpid, parentRpid, mid, uname, avatar, currentLevel, content, ctime]
      properties:
        id: {$ref: '#/components/schemas/PositiveId'}
        rpid: {$ref: '#/components/schemas/PositiveId'}
        parentRpid: {oneOf: [{$ref: '#/components/schemas/PositiveId'}, {type: 'null'}]}
        mid: {type: string}
        uname: {type: [string, 'null']}
        avatar: {type: [string, 'null']}
        currentLevel: {type: [integer, 'null']}
        content: {type: [string, 'null']}
        ctime: {type: string, format: date-time}
    CommentFilter:
      type: object
      properties:
        keyword: {type: [string, 'null'], maxLength: 200}
        mid: {type: [string, 'null'], pattern: '^[1-9][0-9]{0,31}$'}
        uname: {type: [string, 'null'], maxLength: 255}
        levelMin: {type: [integer, 'null'], minimum: 0, maximum: 6}
        levelMax: {type: [integer, 'null'], minimum: 0, maximum: 6}
        unknownLevelOnly: {type: boolean, default: false}
        ctimeFrom: {type: [string, 'null'], format: date-time}
        ctimeBefore: {type: [string, 'null'], format: date-time}
        rpid: {oneOf: [{$ref: '#/components/schemas/PositiveId'}, {type: 'null'}]}
        parentRpid: {oneOf: [{$ref: '#/components/schemas/PositiveId'}, {type: 'null'}]}
        replyScope: {$ref: '#/components/schemas/CommentReplyScope'}
...
    CreateCommentExportRequest:
      type: object
      required: [filter, format]
      properties:
        filter: {$ref: '#/components/schemas/CommentFilter'}
        sort: {$ref: '#/components/schemas/CommentSort'}
        format: {$ref: '#/components/schemas/CommentExportFormat'}
        columns:
          type: array
          minItems: 1
          uniqueItems: true
          items: {$ref: '#/components/schemas/CommentExportColumn'}
    CommentExportJob:
      type: object
      required: [id, taskId, status, format, columns, rowsWritten, bytesWritten, downloadReady, errorCode, errorSummary, createdAt, startedAt, finishedAt, expiresAt]
      properties:
        id: {$ref: '#/components/schemas/PositiveId'}
        taskId: {$ref: '#/components/schemas/PositiveId'}
        status: {$ref: '#/components/schemas/CommentExportStatus'}
        format: {$ref: '#/components/schemas/CommentExportFormat'}
        columns:
          type: array
          minItems: 1
          uniqueItems: true
          items: {$ref: '#/components/schemas/CommentExportColumn'}
        rowsWritten: {type: integer, minimum: 0}
        bytesWritten: {type: integer, minimum: 0}
        downloadReady: {type: boolean}
        errorCode: {type: [string, 'null']}
        errorSummary: {type: [string, 'null']}
        createdAt: {type: string, format: date-time}
        startedAt: {type: [string, 'null'], format: date-time}
        finishedAt: {type: [string, 'null'], format: date-time}
        expiresAt: {type: string, format: date-time}
    CredentialProfile:
      type: object
      required: [id, name, enabled, valid, lastValidatedAt]
      properties:
        id: {$ref: '#/components/schemas/PositiveId'}
        name: {type: string}
        enabled: {type: boolean}
        valid: {type: boolean}
        lastValidatedAt: {type: [string, 'null'], format: date-time}
    ReplaceCredentialSecretRequest:
      type: object
      required: [secret]
      properties:
        secret: {type: string, minLength: 1, maxLength: 16384, writeOnly: true}
...
    CommentPage:
      type: object
      required: [items, nextCursor, total]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/Comment'}}
        nextCursor: {type: [string, 'null']}
        total: {type: integer, minimum: 0}
    ApiProblem:
      type: object
      required: [status, code, detail]
      properties:
        type: {type: [string, 'null'], format: uri}
        title: {type: [string, 'null']}
        status: {type: integer, minimum: 400, maximum: 599}
        code: {type: string}
        detail: {type: string}
        traceId: {type: [string, 'null']}
        fieldErrors: {type: object, additionalProperties: {type: string}}
        existingTaskId: {oneOf: [{$ref: '#/components/schemas/PositiveId'}, {type: 'null'}]}